requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
bokeh==2.3.3
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==1.6.0
cryptography==3.4.7
cssutils==2.3.0
dask==2021.7.0
defusedxml==0.7.1
distributed==2021.7.0
Django==3.2.5
django-allauth==0.45.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-inlinecss==0.3.0
django-timezone-field==4.2.1
dnspython==1.16.0
eventlet==0.31.1
flower==1.0.0
fsspec==2021.7.0
future==0.18.2
greenlet==1.1.0
HeapDict==1.0.1
humanize==3.10.0
idna==3.2
Jinja2==3.0.1
kombu==5.1.0
locket==0.2.1
MarkupSafe==2.0.1
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.0
partd==1.2.0
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psutil==5.8.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pynliner==0.8.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tblib==1.7.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==1.4.4
zict==2.0.0
